Add a bunch of stuff for Mac development.
There are too many differences between Mono and Windows, so I decided to go
with the separate Solution and Project.

This also adds NUnit to the default (Windows) project as well. There's a
post.build.rb Ruby script to copy all of the files over for the noCompile
treatment on Mac.